+++ deezy/audio_encoders/dee/dd.py
@@ -309,6 +309,7 @@ class DDEncoderDEE(BaseDeeAudioEncoder[DolbyDigitalChannels]):
                 cmd=dee_cmd,
                 step_info=step_info,
                 no_progress_bars=self.payload.no_progress_bars,
+                custom_dialnorm=self.payload.custom_dialnorm
             )
         finally:
             self._release_dee()